# services/core-go/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  handler.UpdatePermissionsInput:
    properties:
      permissions:
        items:
          type: string
        type: array
    type: object
...
  response.APIResponse:
    properties:
      data: {}
      errors: {}
      message:
        type: string
      meta:
        $ref: '#/definitions/response.PaginationMeta'
      status:
        type: boolean
    type: object
  response.PaginationMeta:
    properties:
      has_next:
        type: boolean
      has_prev:
        type: boolean
      limit:
        type: integer
      page:
        type: integer
      total_items:
        type: integer
      total_pages:
        type: integer
    type: object
  services.CreateRoleInput:
    properties:
      code:
        type: string
      description:
        type: string
      name:
        type: string
    type: object
...
  services.UpdateRoleInput:
    properties:
      description:
        type: string
      name:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: |-
    High Performance Backend Service for PT. Adijayantara Logistics Indonesia.
    Architecture & Headers:
    - X-API-Version: Target API versioning (e.g. 'v1', 'v2')
    - X-Idempotency-Key: Unique UUID token for safe mutations (POST/PUT/PATCH/DELETE) to prevent duplicates
    - X-Request-ID: Distributed tracing correlation ID
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Cashflow & Shipment Management API
  version: "1.0"
paths:
...
  /roles:
    get:
      description: Retrieves list of all available roles with user counts
      parameters:
      - description: 'API Version (default: v1)'
        in: header
        name: X-API-Version
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: List all roles
      tags:
      - roles
    post:
      consumes:
      - application/json
      description: Creates a new custom business role
      parameters:
      - default: a0eebc99-9c0b-4ef8-bb6d-6bb9bd380a11
        description: Idempotency Key (UUID unik pencegah duplikasi transaksi)
        in: header
        name: X-Idempotency-Key
        type: string
      - description: 'API Version (default: v1)'
        in: header
        name: X-API-Version
        type: string
      - description: New Role Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/services.CreateRoleInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.APIResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Create new custom role
      tags:
      - roles
  /roles/{id}:
    delete:
      description: Deletes a custom role if no users are currently assigned
      parameters:
      - default: a0eebc99-9c0b-4ef8-bb6d-6bb9bd380a11
        description: Idempotency Key (UUID unik pencegah duplikasi transaksi)
        in: header
        name: X-Idempotency-Key
        type: string
      - description: 'API Version (default: v1)'
        in: header
        name: X-API-Version
        type: string
      - description: Role UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Delete custom role
      tags:
      - roles
    get:
      description: Retrieves role information along with active permission codes
      parameters:
      - description: 'API Version (default: v1)'
        in: header
        name: X-API-Version
        type: string
      - description: Role UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Get role detail with active permissions
      tags:
      - roles
    put:
      consumes:
      - application/json
      description: Updates name and description of a role
      parameters:
      - default: a0eebc99-9c0b-4ef8-bb6d-6bb9bd380a11
        description: Idempotency Key (UUID unik pencegah duplikasi transaksi)
        in: header
        name: X-Idempotency-Key
        type: string
      - description: 'API Version (default: v1)'
        in: header
        name: X-API-Version
        type: string
      - description: Role UUID
        in: path
        name: id
        required: true
        type: string
      - description: Updated Role Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/services.UpdateRoleInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Update role information
      tags:
      - roles
  /roles/{id}/permissions:
    put:
      consumes:
      - application/json
      description: Atomically updates the permission checklist assigned to a role
      parameters:
      - default: a0eebc99-9c0b-4ef8-bb6d-6bb9bd380a11
        description: Idempotency Key (UUID unik pencegah duplikasi transaksi)
        in: header
        name: X-Idempotency-Key
        type: string
      - description: 'API Version (default: v1)'
        in: header
        name: X-API-Version
        type: string
      - description: Role UUID
        in: path
        name: id
        required: true
        type: string
      - description: Permission Codes List
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.UpdatePermissionsInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Update role permissions matrix
      tags:
      - roles
  /roles/permissions:
    get:
      description: Retrieves all permissions grouped by modules
      parameters:
      - description: 'API Version (default: v1)'
        in: header
        name: X-API-Version
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: List all available system permissions
      tags:
      - roles
...
securityDefinitions:
  ApiVersion:
    description: Target API Version Routing (e.g. v1, v2)
    in: header
    name: X-API-Version
    type: apiKey
  BearerAuth:
    description: Type "Bearer" followed by a space and the session token.
    in: header
    name: Authorization
    type: apiKey
  IdempotencyKey:
    description: Unique UUID idempotency key for mutation operations (POST/PUT/PATCH/DELETE)
    in: header
    name: X-Idempotency-Key
    type: apiKey
swagger: "2.0"